+++ packages/components/elvia-components.config.js
@@ -336,7 +336,7 @@ module.exports = [
       { name: 'numberOfElements', type: 'number', propType: 'number' },
       { name: 'lastNumberLimit', type: 'number', propType: 'number | undefined' },
       { name: 'dropdownMenuPos', type: 'string', propType: 'string' },
-      { name: 'dropdownItemsDefaultIndex', type: 'number', propType: 'number' },
+      { name: 'selectedDropdownItemIndex', type: 'number', propType: 'number' },
       { name: 'isRightAligned', type: 'boolean', propType: 'boolean | undefined' },
       { name: 'dropdownItems', type: 'object', propType: 'object' },
       { name: 'labelDisplaying', type: 'string', propType: 'string' },